fix: change incident duration from int to float64
- PostgreSQL EXTRACT returns NUMERIC type with decimals
- Go was failing to scan '1.00000000000000000000' into int
- Changed Duration field to float64 to accept decimal values

+++ backend/handlers/public.go
@@ -278,10 +278,10 @@ func (h *PublicHandler) GetServiceUptime(w http.ResponseWriter, r *http.Request)
 	defer rows.Close()
 
 	type IncidentInfo struct {
-		Title       string `json:"title"`
-		Description string `json:"description"`
-		Severity    string `json:"severity"`
-		Duration    int    `json:"duration_minutes"`
+		Title       string  `json:"title"`
+		Description string  `json:"description"`
+		Severity    string  `json:"severity"`
+		Duration    float64 `json:"duration_minutes"`
 	}
 
 	type UptimeDay struct {
